refactor(image): implement image resizing functionality and refactor related code
